refactor(core): extract createActionContextBuilder (PLAN §3.1.f partial)
Pull the buildActionContext + createPerformWithFeedback +
deferHideContextMenu cluster (~170 lines) out of init() into a focused
factory. This is the wiring that bundles together every concern needed
by a plugin action's onAction(context):

  - element + elements (selection)
  - filePath/lineNumber (source)
  - copy() (delegates to copyOrchestrator)
  - enterPromptMode() (delegates to prompt-mode helpers)
  - performWithFeedback() (selection-lock + label + activation lifecycle)
  - hideContextMenu (immediate or deferred)
  - cleanup (deactivate or unfreeze)
  - hooks (plugin transformHtmlContent + onOpenFile + transformOpenFileUrl)

The factory takes 11 inputs covering every cross-cutting concern. The
BuildActionContextOptions interface (was inline in init.tsx) moved to
the controller module.

Drops 5 now-unused imports (normalizeErrorMessage, combineBounds,
PerformWithFeedbackOptions, logRecoverableError, createLabelInstance
+ updateLabelAfterCopy destructures).

index.tsx: 2389 -> 2217 (-172). 46 context-menu + open-file specs green.
